[HK] soft dekete but not running

+++ app/Http/Controllers/Web/Admin/CourseController.php
@@ -152,22 +152,49 @@ class CourseController extends Controller
 
     public function delete($id)
     {
-        $course = Course::findOrFail($id);
-        try {
-            DB::beginTransaction();
+        $course = Course::findorfail($id);
+        $module = DB::table('modules')->where('course_id', $id)->get();
+        $module_rangkuman = DB::table('module_rangkuman')->where('course_id', $id)->get();
+        $quiz = DB::table('quizzes')->where('course_id', $id)->get();
+        $ar = DB::table('augmented_realities')->where('course_id', $id)->get();
 
-            DB::table('modules')->where('course_id', $id)->delete();
-            DB::table('module_rangkuman')->where('course_id', $id)->delete();
-            DB::table('quizzes')->where('course_id', $id)->delete();
-            DB::table('augmented_realities')->where('course_id', $id)->delete();
-            $course->delete();
 
-            DB::commit();
-        } catch (\Throwable $th) {
-            DB::rollBack();
-            return redirect()->route('course.page')->with('error', $th->getMessage());
+        $old_certificate = public_path('storage/images/certificate/') . $course->certificate_course;
+        unlink($old_certificate);
+        if ($course) {
+            $old_image = public_path('storage/images/thumbnail_course/') . $course->image_course;
+            if (file_exists($old_image)) {
+                unlink($old_image);
+            }
+            try {
+                DB::beginTransaction();
+                foreach ($module as $key => $value) {
+                    $old_image = public_path('storage/images/module/') . $value->image_module;
+                    if (file_exists($old_image)) {
+                        unlink($old_image);
+                    }
+                    DB::table('modules')->where('course_id', $id)->delete();
+                }
+                foreach ($module_rangkuman as $key => $value) {
+                    $old_image = public_path('storage/images/module_rangkuman/') . $value->image_module_rangkuman;
+                    if (file_exists($old_image)) {
+                        unlink($old_image);
+                    }
+                    DB::table('module_rangkuman')->where('course_id', $id)->delete();
+                }
+                foreach ($quiz as $key => $value) {
+                    DB::table('quizzes')->where('course_id', $id)->delete();
+                }
+                foreach ($ar as $key => $value) {
+                    DB::table('augmented_realities')->where('course_id', $id)->delete();
+                }
+                $course->delete();
+                DB::commit();
+            } catch (\Throwable $th) {
+                DB::rollBack();
+                return redirect()->route('course.page')->with('error', $th->getMessage());
+            }
         }
-
-        return redirect()->route('course.page')->with('success', 'Kursus telah dihapus');
+        return redirect()->route('course.page')->with('error', 'course telah di hapus');
     }
 }
